move baseline migration to separate file
